[milestone][feat]: Complete FSDR, DSU, GGU implementation for Issue #2
fsdr/ - Feature-Similarity Depth Reuse module (1579 LOC)
- types.py: FSDRConfig, CacheEntry (70-bit), FSDRResult, FSDRProfilingResult
- lsh_hasher.py: LSH signature generation with random projection matrix
- cache_table.py: 128-entry cache with Hamming lookup, LRU+confidence replacement
- depth_corrector.py: Three-level strategy (direct_reuse/interpolation/light_verify)
- light_verifier.py: Local depth search with spread-based range calculation
- fsdr_processor.py: Main engine orchestrating all components
- profiler.py: Hit rate, path distribution, memory savings tracking

dsu/ - Depth Search Unit module (691 LOC)
- types.py: DSUConfig, DSUResult with FSDR statistics
- depth_sampler.py: Projection and bilinear sampling
- cost_volume.py: Cosine similarity cost computation
- softmax_aggregator.py: Probability distribution and statistics extraction
- dsu_processor.py: Full search with FSDR callback creation

ggu/ - Gaussian Generation Unit module (546 LOC)
- types.py: GGUConfig, GaussianOutput (SAES compatible)
- position_calculator.py: 3D position from pixel + depth
- covariance_builder.py: Scale mapping, quaternion to rotation, covariance
- sh_rotator.py: Spherical harmonics rotation (degree 0-3)
- ggu_processor.py: Complete Gaussian generation pipeline

All 141 unit tests passing:
- FSDR: 48/48 passed
- DSU: 20/20 passed
- GGU: 16/16 passed
- SAES: 57/57 passed

Milestone progress: 8670 LOC cumulative, ~92% complete.
Remaining: Adapters and integration module (~765 LOC).

+++ tests/ggu/test_ggu_processor.py
@@ -211,13 +211,13 @@ class TestQuaternionNormalization:
         """Test 90-degree rotation around z-axis."""
         # Quaternion for 90-degree rotation around z: [cos(45°), 0, 0, sin(45°)]
         angle = np.pi / 2
-        quat = torch.tensor([np.cos(angle/2), 0.0, 0.0, np.sin(angle/2)])
+        quat = torch.tensor([np.cos(angle/2), 0.0, 0.0, np.sin(angle/2)], dtype=torch.float32)
         
         R = quaternion_to_rotation_matrix(quat)
         
         # Should map x → y, y → -x
-        x_axis = torch.tensor([1.0, 0.0, 0.0])
-        y_axis = torch.tensor([0.0, 1.0, 0.0])
+        x_axis = torch.tensor([1.0, 0.0, 0.0], dtype=torch.float32)
+        y_axis = torch.tensor([0.0, 1.0, 0.0], dtype=torch.float32)
         
         rotated_x = R @ x_axis
         rotated_y = R @ y_axis
@@ -246,18 +246,18 @@ class TestSHRotator:
         # Degree-1 has 3 coefficients (indices 1, 2, 3)
         sh_1 = torch.tensor([[1.0, 0.0, 0.0],
                             [0.0, 1.0, 0.0],
-                            [0.0, 0.0, 1.0]])  # [3 channels, 3 coeffs]
+                            [0.0, 0.0, 1.0]], dtype=torch.float32)  # [3 channels, 3 coeffs]
         
         # 90-degree rotation around z
         angle = np.pi / 2
-        quat = torch.tensor([np.cos(angle/2), 0.0, 0.0, np.sin(angle/2)])
+        quat = torch.tensor([np.cos(angle/2), 0.0, 0.0, np.sin(angle/2)], dtype=torch.float32)
         R = quaternion_to_rotation_matrix(quat)
         
         # Rotate degree-1 coefficients
         rotated_sh_1 = sh_1 @ R.T
         
         # First channel [1,0,0] should become [0,1,0]
-        assert torch.allclose(rotated_sh_1[0], torch.tensor([0.0, 1.0, 0.0]), atol=1e-5)
+        assert torch.allclose(rotated_sh_1[0], torch.tensor([0.0, 1.0, 0.0], dtype=torch.float32), atol=1e-5)
 
 
 class TestGGUProcessorIntegration:
